[BAEL-3981] Moved inner classes to upper level
