+++ core-java-modules/core-java-lang-2/src/main/java/com/baeldung/comparing/PersonWithoutEquals.java
@@ -1,0 +1,11 @@
+package com.baeldung.comparing;
+
+public class PersonWithoutEquals {
+    private String firstName;
+    private String lastName;
+
+    public PersonWithoutEquals(String firstName, String lastName) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+}
